% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection_method.R
\name{bisection_method}
\alias{bisection_method}
\title{Bisection method}
\usage{
bisection_method(f, x0, x1, tol = 1e-08)
}
\arguments{
\item{f}{Univariate function to find root of}

\item{x0}{A point}

\item{x1}{A point larger than \code{x0}}

\item{tol}{Tolerance for convergence.}
}
\value{
A root of \code{f}. If the algorithm does not converge, \code{NA}
is returned.
}
\description{
The bisection method is an iterative root-finding method that does not use
derivatives.
}
\details{
The bisection method finds the root of a univariate function \eqn{f}
given two initial guesses \eqn{x_0} and \eqn{x_1} where \eqn{x_0 < x_1} and
\eqn{sign(f(x_0)) = -sign(f(x_1))} by sequentially bisecting the interval
(\eqn{x_2 = (x_0 + x_1) / 2}).

The algorithm terminates when:
\itemize{
\item the algorithm exceeds 1000 iterations,
\item the value of \eqn{f} is non-finite for an iterate (\eqn{x_n}),
\item the algorithm converges and \eqn{|f(x_{n}) - f(x_{n + 1})| < tol}.
}
}
\examples{
f <- cos
x0 <- 0
x1 <- pi
bisection_method(f, x0, x1)

}
