% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halleys_method.R
\name{halleys_method}
\alias{halleys_method}
\title{Halley's method}
\usage{
halleys_method(f, fp, fpp, x0, tol = 1e-08)
}
\arguments{
\item{f}{Univariate function to find root of}

\item{fp}{First derivative of \code{f}}

\item{fpp}{Second derivative of \code{f}}

\item{x0}{A point close to the root of \code{f}}

\item{tol}{Tolerance for convergence.}
}
\value{
A root of \code{f} near \code{x0}. If the algorithm does not
converge, \code{NA} is returned.
}
\description{
Halley's method is an iterative root-finding method with cubic
convergence that requires the first and second derivative.
}
\details{
Halley's method finds the root of a univariate function \eqn{f}
with first derivative \eqn{f'} and second derivative \eqn{f''} given an
initial guess \eqn{x_0} by the iteration:
\deqn{x_{n + 1} = x_n - \frac{2f(x_n)f'(x_n)}{2(f'(x_n))^2 - f(x_n)f''(x_n)}}.

The algorithm terminates when:
\itemize{
\item the algorithm exceeds 1000 iterations,
\item the value of \eqn{f}, \eqn{f'}, or \eqn{f''} is non-finite for an
iterate (\eqn{x_n}),
\item the iterate (\eqn{x_n}) becomes non-finite, or
\item the algorithm converges and \eqn{|f(x_{n}) - f(x_{n + 1})| < tol}.
}
}
