% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtons_method.R
\name{newtons_method}
\alias{newtons_method}
\title{Newtons's method}
\usage{
newtons_method(f, fp, x0, tol = 1e-08)
}
\arguments{
\item{f}{Univariate function to find root of}

\item{fp}{First derivative of \code{f}}

\item{x0}{A point close to the root of \code{f}}

\item{tol}{Tolerance for convergence.}
}
\value{
A root of \code{f} near \code{x0}. If the algorithm does not
converge, \code{NA} is returned.
}
\description{
Newtons's method is an iterative root-finding method with quadratic
convergence that requires the first derivative.
}
\details{
Newtons's method finds the root of a univariate function \eqn{f}
with first derivative \eqn{f'} given an initial guess \eqn{x_0} by the
iteration:
\deqn{x_{n + 1} = x_n - \frac{f(x_n)}{f'(x_n)}}.

The algorithm terminates when:
\itemize{
\item the algorithm exceeds 1000 iterations,
\item the value of \eqn{f} or \eqn{f'} is non-finite for an iterate
(\eqn{x_n}),
\item the iterate (\eqn{x_n}) becomes non-finite, or
\item the algorithm converges and \eqn{|f(x_{n}) - f(x_{n + 1})| < tol}.
}
}
\examples{
newtons_method(cos,
               function(x) -sin(x),
               0.5)
newtons_method(function(x) x ^ 3 - x - 2,
               function(x) 3 * x ^ 2 - 1,
               1)

}
