% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secant_method.R
\name{secant_method}
\alias{secant_method}
\title{Secant method}
\usage{
secant_method(f, x0, x1, tol = 1e-08)
}
\arguments{
\item{f}{Univariate function to find root of}

\item{x0}{A point close to the root of \code{f}}

\item{x1}{A point close to the root of \code{f} not equal to \code{x0}}

\item{tol}{Tolerance for convergence.}
}
\value{
A root of \code{f}. If the algorithm does not converge, \code{NA}
is returned.
}
\description{
The secant method is an iterative root-finding method with order of
convergence phi (1.618...) that does not use derivatives.
}
\details{
The secant method finds the root of a univariate function \eqn{f}
given two initial guesses \eqn{x_0} and \eqn{x_1} by the
iteration:
\deqn{x_{n + 1} = x_n - f(x_n) \frac{x_n - x_{n - 1}}{f(x_n) - f(x_{n - 1})}}.

The algorithm terminates when:
\itemize{
\item the algorithm exceeds 1000 iterations,
\item the value of \eqn{f} is non-finite for an iterate (\eqn{x_n}),
\item the iterate (\eqn{x_n}) becomes non-finite, or
\item the algorithm converges and \eqn{|f(x_{n}) - f(x_{n + 1})| < tol}.
}
}
\examples{
secant_method(cos, 0, 1)
secant_method(function(x) x ^ 3 - x - 2, 1, 2)

}
