% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steffensens_method.R
\name{steffensens_method}
\alias{steffensens_method}
\title{Steffensen's method}
\usage{
steffensens_method(f, x0, tol = 1e-08)
}
\arguments{
\item{f}{Univariate function to find root of}

\item{x0}{A point close to the root of \code{f}}

\item{tol}{Tolerance for convergence.}
}
\value{
A root of \code{f} near \code{x0}. If the algorithm does not
converge, \code{NA} is returned.
}
\description{
Steffensen's method is an iterative root-finding method with quadratic
convergence that does not use derivatives.
}
\details{
Steffensen's method finds the root of a univariate function \eqn{f}
given an initial guess \eqn{x_0} by the iteration:
\deqn{x_{n + 1} = x_n - \frac{f(x_n)}{g(x_n)}}
where
\deqn{g(x_n) = \frac{f(x_n + f(x_n))}{f(x_n)} - 1}.

The algorithm terminates when:
\itemize{
\item the algorithm exceeds 1000 iterations,
\item the value of \code{f} is non-finite for an iterate (\eqn{x_n}),
\item the iterate (\eqn{x_n}) becomes non-finite, or
\item the algorithm converges and \eqn{|f(x_{n}) - f(x_{n + 1})| < tol}.
}
}
