% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Univariate Plot}
\usage{
univariate(frame, groupvar, actual, model, grouprange = NULL,
  roundvalue = NULL, roundfunc = "round", labelangle = 0,
  titlestring = NULL)
}
\arguments{
\item{frame}{Dataframe, containing groupvar, actual, predict as columns}

\item{groupvar}{Variable to aggregate over}

\item{actual}{Observed values}

\item{model}{Predicted values}

\item{grouprange}{Range to plot groupvar over. Defaults to full range}

\item{roundvalue}{Optional rounding accuracy for groupvar. Defaults to no rounding}

\item{roundfunc}{Rounding function (round, floor, ceiling). Defaults to round}

\item{labelangle}{Degrees from zero to display X-axis labels}

\item{titlestring}{String for plot title. Default title used if NULL}
}
\value{
A ggplot object
}
\description{
Plot predicted and actual values over an aggregating variable, with histogram
}
\examples{
univariate(train, "SalesBranch", "STPFlag", "predict")
univariate(df, "InterestRate", "FundedFlag", "predict", c(2, 4.5))
univariate(test, "FICO", "FundedFlag", "predict", c(580, 850), roundvalue=5)

}
