% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarGraphs.R
\name{univariateCategorical}
\alias{univariateCategorical}
\title{A univariate graph function}
\usage{
univariateCategorical(ip.data, cat_var, plot.name)
}
\arguments{
\item{ip.data}{is your dataframe with data}

\item{cat_var}{is the column name of your categorical variable}

\item{plot.name}{the name of the plot u want to see where a = bar chart
b = horizontal bar , c = pie chart , d = point}
}
\description{
A univariate function to generate graphs for categorical variables
}
\examples{
univariateCategorical(iris,iris$Species,"a")

}
\keyword{A}
\keyword{basic}
\keyword{be}
\keyword{function}
\keyword{in}
\keyword{plot}
\keyword{returned}
\keyword{this}
\keyword{will}
