% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarGraphs.R
\name{univariateNumerical}
\alias{univariateNumerical}
\title{A univariate graph function}
\usage{
univariateNumerical(ip.data, num_var, plot.name)
}
\arguments{
\item{ip.data}{is your dataframe with data}

\item{plot.name}{the name of the plot u want to see where a = density plot
b = box plot , c = histogram , d = jitter}

\item{cat_var}{is the column name of your numerical variable}
}
\description{
A univariate function to generate graphs for Numerical variables
}
\examples{
univariateNumerical(mtcars,mtcars$mpg,"a")

}
\keyword{A}
\keyword{basic}
\keyword{be}
\keyword{function}
\keyword{in}
\keyword{plot}
\keyword{returned}
\keyword{this}
\keyword{will}
